backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
httpx==0.27.0
sqlalchemy>=2.0
asyncpg>=0.29
openai>=1.40.0
python-dotenv